requirements.txt
Django>=5.0,<6.0
psycopg[binary]>=3.2
python-dotenv>=1.0
Pillow>=10.0
django-crispy-forms>=2.1
crispy-bootstrap5>=0.7
pandas>=2.2
openpyxl>=3.1
